+++ src/lib.rs
@@ -1,2 +1,21 @@
+use leptos::ev::MouseEvent;
+
 pub mod components;
 pub mod elements;
+
+pub struct MouseEventFn(Box<dyn Fn(MouseEvent) + 'static>);
+
+impl<T> From<T> for MouseEventFn
+where
+    T: Fn(MouseEvent) + 'static,
+{
+    fn from(value: T) -> Self {
+        Self(Box::new(value))
+    }
+}
+
+impl MouseEventFn {
+    pub fn into_inner(self) -> Box<dyn Fn(MouseEvent) + 'static> {
+        self.0
+    }
+}